TVGUI-17 Move Reminders components into components sub-directory
