--- src/Reminders/ReminderCard.css
@@ -1,16 +1,0 @@
-.reminder-card {
-    border: solid 1px black;
-    background-color: white;
-    color: black;
-    max-width: 50%;
-    margin: 0 auto;
-    transition: transform 2s ease;
-    margin-top: 1%;
-    margin-bottom: 1%;
-}
-
-.reminder-card:hover {
-    background-color: black;
-    color: white;
-    transform: scale(1.2);
-}
