+++ src/components/Reminders/EditReminder.tsx
@@ -1,9 +1,9 @@
 import { useState, useContext, FormEvent, SetStateAction, Dispatch } from "react";
 import { Alert, Form, Input, Modal, Select } from "antd";
 
-import { editReminder } from "../requests/requests";
-import { UserContext } from "../contexts/UserContext";
-import { Reminder } from "../utils";
+import { editReminder } from "../../requests/requests";
+import { UserContext } from "../../contexts/UserContext";
+import { Reminder } from "../../utils";
 
 interface EditReminderProps {
     reminderChosen: Reminder
