+++ src/components/Reminders/index.ts
@@ -1,0 +1,4 @@
+import AddReminder from "./AddReminder";
+import EditReminder from "./EditReminder";
+
+export { AddReminder, EditReminder };
